+++ packages/utils/src/index.ts
@@ -1,4 +1,4 @@
 export { useOutsideClick } from './hooks/useOutsideClick';
 export { useOverlay } from './hooks/useOverlay';
 export { getTabLabel } from './util/tab';
-export { isAfterMonth, isDateDisabled, isDateToday, isDateSelected, getDayVariant } from './util/datePicker';
+export { isAfterMonth, isDateDisabled, isDateToday, isDateSelected, getDayVariant, generateCalendarData } from './util/datePicker';
